+++ instrumentation/opentelemetry-instrumentation-asyncua/src/opentelemetry/instrumentation/asyncua/__init__.py
@@ -1,0 +1,5 @@
+from opentelemetry.instrumentation.asyncua.instrumentor import (
+    AsyncUAInstrumentor,
+)
+
+__all__ = ["AsyncUAInstrumentor"]